style(command): 修改目录
修改目录
