+++ sky-axon-command/src/main/java/com/sky/axon/command/controller/AccountQueryController.java
@@ -20,16 +20,36 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package com.sky.axon.web.core.command;
+package com.sky.axon.command.controller;
+
+import com.sky.axon.api.query.AccountQueryDTO;
+import com.sky.axon.query.model.Account;
+import com.sky.axon.query.service.AccountQueryService;
+import io.swagger.annotations.Api;
+import org.springframework.web.bind.annotation.*;
+
+import javax.annotation.Resource;
+import java.util.List;
 
 /**
- * 删除账号
- *
  * @author
  */
-public class RemoveAccountCommand extends BaseCommand<String> {
+@RestController
+@RequestMapping(value = "/bank-accounts")
+@Api(value = "Account Queries", description = "Account Query Events Endpoint", tags = "Account Queries")
+public class AccountQueryController {
 
-    public RemoveAccountCommand(String id) {
-        super(id);
+    @Resource
+    private AccountQueryService accountQueryService;
+
+
+    @GetMapping("/findEvents")
+    public List<Object> findEvents(@RequestParam("id") String id) {
+        return accountQueryService.listEventsForAccount(id);
+    }
+
+    @GetMapping("/findAccount")
+    public List<Account> findAccount(@RequestBody AccountQueryDTO accountQueryDTO) {
+        return accountQueryService.findAccount(accountQueryDTO);
     }
 }
