+++ sky-axon-command/src/resources/application.properties
